+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Fixed
 - Fix a bug where we would re-use the same stream on a failed upload, instead of a new one.
+- Fix a bug where deleting a cloud video would not trigger other clients to update.
 
 ## [5.3.1] - 2024-05-08
 ### Fixed